Drop modifications to libsharp
I do not believe we are allowed to integrate modified GPL code into
the SpECTRE source.  There may be an exception if the modified code
could still be used separately, but the changes we made to libsharp
did not allow that since we replaced its build system with SpECTRE's.

This replaces the contents of the external/libsharp directory with the
unmodified source, and builds it as an external project.